+++ src/services/questions.js
@@ -1,5 +1,8 @@
-import questions from '../question.json'
+// import questions from '../question.json';
+import { fetchAllQuestions } from "./firebase";
+
 
 export const getData = () => {
-    return Promise.resolve(questions);
+    // return Promise.resolve(questions);
+    return fetchAllQuestions();
 }
